Introduce get_power_state for drivers
Also, provide a KeepaliveLibvirtDriver as an alternative to
LibvirtDriver to stay connected to the server at all times. This mode
will not correctly (read: not at all) handle reconnects, and should not
be used for now. This mode allows integration tests to avoid using the
test+unix mode. Using the latter is dangerous because of race conditions
when running multiple tests at the same time. In fact, libvirt launches
a (single) daemon for the whole duration of the process. This means
multiple tests accessing the same libvirtd instance, causing all kinds
of errors.