Ensure APCRackPDU instanciation does not update core
Currently, APCRackPDU will notify the core for all of it's port being
updated as power state "ON" because of improper usage of the PDU class.
In fact, the PDU class provides support for "default" values that bypass
the callback.

This commit uses the right field and adds a unit test to cover that
case.