Support listen_address in config instead of listen_host
listen_address was the intended name but listen_host was written in the
tests, which led to us read from listen_host, but use listen_address
throughout the code.

Solve this by reading listen_address only. Backwards compatibility is
not a concern as this is pre-release software, and there are no official
uses of it yet.

Closes #19