Move the outlet state responsibility to the core
Now, the core stores the outlet state and the SNMP PDU adapter only
notifies and consult the core about the state of an outlet.  In the
case where the outlet has never been set, the driver is queried to get
the outlet state.  In the event that there is no device (libvirt
domain) associated to an outlet, the default outlet state from the
configuration will be returned.

Also, the outlet default state is now in the "global" section of the
configuration as there is no outlet defalut state per PDU anymore.

    [global]
    outlet_default_state = ON