Let PDU outlets notify the core directly
In order to remove the mutual dependency of the PDU and the PDUOutlet
class, the PDUOutlet class will be responsible to talk to the core and
make the appropriate data transformations.

This will allow the PDU class to, eventually, manage more OID registers
associated to the outlet than simply its state.