Add support for custom PDU outlet count
The number of outlet may now be specified at the constructor call of the
PDU class.  If defined, it will override the PDU class specific outlet
count.

Also, the SNMPPDUHarness.stop method was not protected with a lock.
Without this, the stop method simply fails to stop the thread if called
just after the start method.

Reviewed-By: Maxime Dupuis <mdupuis@internap.com>